requirements.txt
Django==4.2.21
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.3.1
django-filter==23.3
psycopg[binary]==3.2.3
celery==5.3.1
redis==4.6.0
Pillow==11.0.0
python-decouple==3.8